pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-dependency-analyzer"
version = "0.1.0"
description = "Dependency network analyzer for the ML ecosystem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "ratelimit>=2.2.1",
    "tomli>=2.0.1",
    "requirements-parser>=0.5.0",
    "networkx>=3.2",
    "pandas>=2.1.0",
    "streamlit>=1.29.0",
    "pyvis>=0.3.2",
    "plotly>=5.18.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
